+++ hedera-mirror-rosetta/app/services/construction_service.go
@@ -205,7 +205,7 @@ func (c *ConstructionAPIService) handleCryptoTransferPayload(operations []*rType
 			return nil, errors.Errors[errors.InvalidAccount]
 		}
 
-		amount, err := parse.ToInt(operation.Amount.Value)
+		amount, err := parse.ToInt64(operation.Amount.Value)
 		if err != nil {
 			return nil, errors.Errors[errors.InvalidAmount]
 		}
