+++ hedera-mirror-rosetta/tools/parse/parse.go
@@ -2,6 +2,6 @@ package parse
 
 import "strconv"
 
-func ToInt(value string) (int64, error) {
+func ToInt64(value string) (int64, error) {
 	return strconv.ParseInt(value, 10, 64)
 }
