+++ hedera-mirror-rosetta/tools/validator/validator.go
@@ -9,7 +9,7 @@ import (
 func ValidateOperationsSum(operations []*types.Operation) *types.Error {
 	var sum int64 = 0
 	for _, operation := range operations {
-		amount, err := parse.ToInt(operation.Amount.Value)
+		amount, err := parse.ToInt64(operation.Amount.Value)
 		if err != nil {
 			return errors.Errors[errors.InvalidAmount]
 		}
